feat: Rename template to OpenResume
