+++ .github/workflows/deploy.yaml
@@ -31,7 +31,7 @@ jobs:
       - name: Install dependencies
         run: pnpm install
       - name: Build
-        run: pnpm run build -- --preset github_pages
+        run: pnpm run build --preset github_pages
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
