+++ globals/index.ts
@@ -4,5 +4,5 @@ export const templates: Array<Template> = [
   "Benjent2023",
   "Benjent2023Light",
   "Caillou2023Light",
-  "Openresume2023Light",
+  "OpenResume",
 ];
